Archive: Import new tweets table name
Given the code here, it seems like in the past this file was named
"tweet.js".  In recent exports, it's named "tweets.js".  The archive
importer needs to be modified to take this into account.  Existing logic
is reused for importing this table.  (However, the resulting table name
will be different, matching the different file name -- archive_tweets,
rather than archive_tweet).